requirements.txt
chemcrow
openai
serpapi
rdkit-pypi
python-dotenv
paperscraper
numpy<2.0
pydantic==1.10.13
langchain==0.0.275
numexpr==2.8.4
